Update patch set 8
Patch Set 8: Code-Review-1

> I wouldn't check for content.  I'd check that only one exists and fail to mount anything if there are multiple mount points prepared.

IMHO multiple directories setup _blank_ should not be considered an error. The error should be if multiple entries have override data in the upper directory.

It is overly complex to determine at late runtime (adb disable-verity) which mount is the best candidate for timing, so this would require the bringup engineer to be burdened with coming up with an answer to this question and then setting an ro.adbd.mount.overlayfs property for the product. A burden that can be answered _wrong_ since it is a direct derivative of any adjustment(s) made in the device tree and fstab. Making an engineer have to remember to adjust the ro.adbd.mount.overlayfs property to match the fstab set adjustments is prone to error. Change the kernel, decisions change, answer will undoubtedly be wrong in ro.adbd.mount.overlayfs ...

It is not necessary to add it until they want/need to specify it, especially if a less than optimal option needs to be taken (eg: /cache is there, but /data is preferred even if late mounted). IMHO populating all the trees must be an option at runtime, boot time is imminently capable of deciding which one to choose based on first mount (set) available.

A persist.adb.mount.overlayfs option can certainly can contribute to the decision if the user prefers, but there is no harm in the any-or-all case since at boot time it selects the _best_ and _earliest_ one always unless otherwise selected by the properties. We may wish to consider adding an option to adb remount to add the ability to select one?

The CL to add ro.adbd.mount.overlayfs is where these options can be added, and I would like to plead that this later CL refinement is a better place to work this out. The automated case will provides for the squashfs Hikey, A/B and non A/B devices without the need for an engineer to fondle a tuneable.

That said, I am still doing the selection loop wrong in this version of the CL, priority should be instead on which _ONE_ has existing override data to stick with that one. If none has override data, followed by first with boot time R/W capability success check for workdir. Yes, an error and do not mount, if multiple have upperdir content.

Patch-set: 8
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1