Update patch set 16
Patch Set 16:

(16 comments)

A lot of style comments that will significantly help readability/reviewability inline.

I see in the comments that the plan is to use the classes to rewrite engine.cpp.  Do you have a CL that does that, even if WIP?  Seeing the end usage of the classes will help understand how this all fits together.

It doesn't seem like there is a good match between the actions in engine.cpp and the Command classes.  If the only user of these classes is the FastBootDriver class, then it would be better if those classes were just coded directly as functions.

What is the difference between the new FastBootDriver class and the existing functions in protocol.cpp?  These seem to have a lot of overlap and should be merged if any changes are to be done.

Patch-set: 16
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>