Update patch set 14
Patch Set 14:

> Are you supposing that the underlying mechanism will be fixed to system properties and won't change forever? If that's true, yes we need neither additional accessibility checks nor any classes. 

I don't foresee this happening in the near future.  

Regardless though, the new underlying mechanism needs real enforcement of the accessibility requirements (something like SELinux that will prevent the improper access) and not just a client side check as this is done now.  And with such enforcement in place, these client side checks are superfluous.

> Even better, we can just add new getters and setters with various types to libbase (like android::base::GetIntProperty). But I wanted to remove direct dependency between this library and System Properties, and leave the possibility to change the underlying mechanism without breaking any interfaces, even if it's unlikely to occur.

That's fine; I agree with that goal.  But the actual API exposure shouldn't contain any of this.  It should just be getters and setters which call into the underlying APIs.  In the properties case it can be as simple as just forwarding to Get/SetProperty() calls and handling the additional parsing.

> For example, we can allow a sysprop to be accessible from platform libraries while disallowing it from vendor libraries even when the libraries are loaded in the same process.

As written above, you can't though unless you have an enforcement mechanism for this.

> I think that's more or less a matter of style. For example, the Java counterpart of this (ag/710826) is designed in this way and people liked it. I would like to keep the symmetry across C++ and Java APIs.

I don't think that's a necessary goal as C++ is a fundamentally different language than Java.  There is going to be a ton of static constructor overhead as well as memory overhead to build classes and the interface is not any clearer.

> The PropertyNamespace in the CL is can be thought as the kind of the owner. This will of course be encoded into the namespace (or package) of the API.

But why include it as a data member?  When is this used, particularly by the client?  It can't be used for access checks as there's no way of enforcing it on the client side, so it seems like it's just extra data hanging around without being used.

Patch-set: 14
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>