Update patch set 14
Patch Set 14:

> 1) The Accessor functions, PlatformWritable(), etc are problematic.
I don't intend to replace or eliminate the access control by the selinux. But in some cases, they can do what can't be done merely using selinux. For example, we can allow a sysprop to be accessible from platform libraries while disallowing it from vendor libraries even when the libraries are loaded in the same process.

> 2) There is no reason to keep the concept of 'write once'.
ok.

> 3) There is no reason for these to be classes.
I think that's more or less a matter of style. For example, the Java counterpart of this (ag/710826) is designed in this way and people liked it. I would like to keep the symmetry across C++ and Java APIs.
BTW, these are just base classes. The actual API will be generated from the sysprop description and will be plain simple functions for C++. The signature will be similar to what you wrote.

> 4) I'm a bit iffy about the idea of the namespaces.
We do encode the namespace of a sysprop into the API for it. That part is not shown in this CL which implements only the base classes. The generated code will be under appropriate namespaces (or package in Java) Inseob is working on that too.

The PropertyNamespace in the CL is can be thought as the kind of the owner. This will of course be encoded into the namespace (or package) of the API.

Patch-set: 14
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>