Update patch set 14
Patch Set 14:

> However, 1) is particularly important and is the main reason for having the classes.

Isn't 1) available during compile time?  You know when a module is being compiled if it's a platform or vendor module.  You know the access bits for each property.  Even if you don't do this in the headers, can't you create a 'platform system properties' library with symbols for all of the properties that the platform can access and a 'vendor system properties' library with symbols for all of the properties that vendor modules can access.  Then you'd restrict system processes to only link the former and vendor processes to only link the latter.  In any case, if you don't go with the header only idea, then you're doing to need to have symbols for literally every property anyway, so it's not like this is any overhead.  And if you do this, you get compile time errors and minimize runtime overhead.

> P.s. we don't yet actually hide the old sysprop APIs (property_get, etc.) But the plan is to make the APIs visible only to this library (or callable only from this library).

How do you plan to do this?  enh@ mentioned this offline, but it's an NDK API, so it isn't exactly trivial.

Patch-set: 14
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>