Update patch set 22
Patch Set 22:

(2 comments)

Something pointed out to me today; does this break the windows / mac transports or does this still handle them?

I'm still not convinced this is a change we should take.  This is essentially a replacement for protocol.cpp, but I'm not convinced that this is overall any better.

For your purposes, what was the old code lacking?  Don't you really just need an ability to send and receive raw commands which protocol.cpp already had?

Patch-set: 22
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>