Update patch set 22
Patch Set 22:

(2 comments)

> Patch Set 22:
> 
> (2 comments)
> 
> Something pointed out to me today; does this break the windows / mac transports or does this still handle them?
> 
> I'm still not convinced this is a change we should take.  This is essentially a replacement for protocol.cpp, but I'm not convinced that this is overall any better.
> 
> For your purposes, what was the old code lacking?  Don't you really just need an ability to send and receive raw commands which protocol.cpp already had?

> Something pointed out to me today; does this break the windows / mac transports or does this still handle them?

Why would this be the case? Transport is an ABC where the appropriate derived class is managed and instantiated by the caller (same as before), and passed in to the constructor. It is used exactly the same as before.

> I'm still not convinced this is a change we should take.  This is essentially a replacement for protocol.cpp, but I'm not convinced that this is overall any better.

Please see the design doc (https://docs.google.com/document/d/1RvA6SWc3B8sLtjn-EvdQSk2Q3ngS0mgdOl73EZXkUbI/edit?usp=sharing). Specifically, take a look at the Background section. Please note that most of the design doc is old and needs to be updated, but the background section is still relevant.

> For your purposes, what was the old code lacking?  Don't you really just need an ability to send and receive raw commands which protocol.cpp already had?

I think if you compared what is currently fastboot_driver.cpp with the old protocol.cpp, you will find they are very similar. The only difference is the deficiencies outlined in the design doc are addressed and the code is overall cleaned up, made OO, and modified to use safer C++ data structures.

Patch-set: 22
Reviewer: Gerrit User 1336956 <1336956@85c56323-6fa9-3386-8a01-6480fb634889>