Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> (I'm about the abandon this CL.)
> 
> The fs_mgr change has been merged via https://android-review.googlesource.com/c/platform/system/core/+/721667.
> 
> The init change doesn't make real sense, as this CL changes the init_defaults rules, but not the actual init executable (which is still statically linked). And because init is not built with Android.bp, we can't safely switch that to using shared libs (some of the depended shared libs may not be copied to recovery).
> 
> I'm making the adb related changes in a separate CL, which is currently WIP. But it shouldn't block other CLs.

Okay, thanks for the information. Unfortunately fastbootd fails to link without fixes to libadbd, because it's missing those services. I'd like to keep the fastbootd CLs building in the meantime, so I'll pare down this to the bare minimum, -2 it, and abandon it as soon as your other fixes are in.

Patch-set: 10
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>