Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What is the plan with internal vs external lmkd?  Is the internal version with statsd going to be the final version?  In which case is this needed?
> 
> How are these APIs used currently?  Is it that you can only ever call android_log_write_list() once per create_android_logger() call and the goal here is to be able to reuse this result?  What is stopping us from having android_log_write_list() clear the data that it has written out from the context, effectively performing this reset automatically?
> 
> Otherwise this is fine, just wondering the above.

Internal lmkd with statsd is used for the pi-dev release and I assume this will be the case with the future releases.

I believe the difference between AOSP and internal was created because some pieces of statsd were not available in AOSP and Android Go team needed lmkd stats collection (Mark should know more about this and what exactly is missing in AOSP). I would love to resolve this difference somehow because it creates lots of merge conflicts but not sure how. Maybe Mark can help me with the answer.

I think this change is still valid and the implementation in libstats hopefully can be replaced with this one.

As for doing automatic resets in android_log_write_list() I'll let Mark as the owner answer this question.

Patch-set: 2
Reviewer: Gerrit User 1157738 <1157738@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0