Update patch set 5
Patch Set 5:

On a second thought, it looks cleaner to drop libsync_recovery with a separate CL, since this one also contains the changes to libadf. I'll revert this change to PS4 and land separately. The change that merges libsync_recovery and libsync is in https://android-review.googlesource.com/c/platform/system/core/+/727312.

Patch-set: 5