Update patch set 17
Patch Set 17: Code-Review+2

(1 comment)

> I've removed the thread for now. It seems unsafe to have this going on in the background while new commands come in on the main thread.

Makes sense to me.  The thread was a good idea, but I think the overall fastboot protocol is meant to be synchronous, so there's unfortunately not a way to defer the FAIL/OKAY status of a command while others are run and dropping it is best.

Patch-set: 17
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2