Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> Ok for getting a first version in, but should we wait for system/apexd to be created? I can anyway work on the binder changes for init independently.

We can merge here first and move it to other place when the git is created. I've requested the creation of /system/apex (not apexd) to host apexd, packager, and other libs/tools that we might need in the future.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>