Update patch set 8
Patch Set 8:

Btw, for current A/B devices (in N, O, P), usually /boot is re-purposed for "recovery" and there is no /recovery partition.

We are now replacing recovery-ramdisk in /boot with first-stage-ramdisk, which seems means A/B legacy devices might not be able to switch to first-stage-ramdisk device? (this shouldn't be an issue though).

Patch-set: 8
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>