Update patch set 1
Patch Set 1:

notwithstanding the _real_ fix for rename, I agree.

Will spin a CL to change bool to int, errno stuffed into negative values of int. Parameter *change remains to cover error and something changed.

The messages of concern are associated with PLOG, and for cases errno is acceptable. The mechanics of PLOG trickery bother me so I will admittedly think about this. However we must admit that adb has an if (errno) PLOG fragment that is fishy enough to require at least return 'can do it', 'will not do it', 'there was an error doing it and here it is' and 'something was changed while doing it, error or no'.

Do you want to consider a way to pass in an errno value to PLOG-like operations (or is there already one)?

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0