Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Is the reboot-fastboot command here sufficient in that case?  Presumably we'd want an explicit set of commands to allow rebooting from either bootloader or userspace fastboot into each of them.
> 
> If I'm understanding correctly: you're thinking of a case where the device is in userspace fastboot, but the host fastboot tool needs to do something that can only be done through the bootloader? 

Well, yes, if there are non-HLOS commands that only the bootloader fastboot can do.  Perhaps this is just something that we need to work with OEMs for?  Or even before then, we should figure out if we can do the full range of fastboot commands currently with userspace fastboot.  We don't support locking / unlocking yet, right?  Are those going to be easy to support or are some of those keys only baked into / accessible by the bootloader?

Hridya, what do you think? Is there a scenario where the HAL could say, "please reboot to bootloader to do this", and we'd want the host fastboot tool to automatically do so?
> 
> Slightly related, the last change I'd like to make here is that boot partitions get flashed first, before potentially rebooting. Otherwise if you're stuck in bootloader fastboot you can't use "flashall" at all. So the new order of operations would be:
> 
> (1) Flash boot, vbmeta, recovery, dts, dtbo.
> (2) Reboot to userspace if needed.
> (3) Flash everything else.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>