Update patch set 4
Patch Set 4: Verified+1 Presubmit-Ready+1

(3 comments)

> Without some strong motivating cases i'd suggest trying to keep the behavior similar to the existing (especially some of the more subtle changes like dropping empty values)

Agreed.  I was wanting to limit this to something more sane, but in the end I'm going to keep it just as flexible as it was.  Just this time with regexes and tests + the rest of the clean up which is what I actually wanted to do.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1