Update patch set 3
Patch Set 3: Code-Review+2

This isn't a compiler bug. There is an implicit sign conversion for 0xFF, which is an integer with value 255, and then later a use of it as a signed character (which 255 cannot be represented by). The use of integer sanitization (part of UBSan) and the presence of "-fno-sanitize-recover=integer" is what lets the compiler fold away all the invalid conversions/comparisons. If I remove that flag, the comparisons come back. ndesaulniers@ points out that we should really make sure that -Wsign-conversion is required in the Android platform (it isn't part of -Wall).

Patch-set: 3
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2