Update patch set 5
Patch Set 5:

(1 comment)

> '/' is showing up on one of the devices, and it is getting incorrectly handled.

Which device?  Is it system-as-root?

This is happening specifically during first stage mount?  '/' should show up once the system is booted and the full fstab is available, but it shouldn't be present in the first stage fstab (currently at least; we could think about changing that too).

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>