Update patch set 5
Patch Set 5:

At the time that fs_mgr_overlayfs_* is called, the default fstab is available (on /vendor/etc/fstab.<hardware>.rc for example) to be read, so '/' could exist on those systems.

The good news is that update_verity_state gives us the much needed report of "vroot" during first stage init at this point, whereas it was marking /system as not verity protected even though it was and allowing overlayfs to mount (if the backing directory existed, which if everything is going as planned, it does not, but I _can_ not allow that because of the strict security requirements I was handed).

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>