Update patch set 4
Patch Set 4:

Not to re-neg on my previous +2, but from the conversation today, I think we concluded that:

1) This code is called after vendor is mounted, thus we have both the vendor fstab and the DT fstab as the default fstab
2) Either /system will exist in the DT fstab or '/' will exist in the vendor fstab.

Unless we know of a specific case where 2) is not true, then we can likely drop this recovery code.  Alternatively, we can drop this recovery code anyway and make it a requirement that system-as-root device need to specify '/' in their fstab in order for overlayfs to work.

Point being that if we do that, we can do what I've been hoping to do this entire time, and have a helper function for fs_mgr_update_verity_state() that takes an fstab_rec as an input and doesn't log.  It'll simplify the code here and it'll make https://android-review.googlesource.com/c/platform/system/core/+/784636 unneeded.

Patch-set: 4