Update patch set 4
Patch Set 4:

Okay, close enough.  One day we'll get this refactored sanely.

If we do move to having an fstab on the ramdisk (or recovery) that is identical to what is on /vendor, then we'll at least know we'll always have the same 'default' fstab in all locations, which might sell me on doing that sooner than later, just to clean up this mess of not knowing which fstab is doing what.

Patch-set: 4