Update patch set 1
Patch Set 1:

Ah, the problem is that we have assertions that things are only run on the main thread, and some tests were missing calls to set the main thread. When run with all of the tests, it'd pass because a previous test would make the call, and when run with just the individual test, it'd just work. Uploading a CL to fix now.

Patch-set: 1