Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> If I understand correctly, the main question here is whether overlapping libraries are OK. Martin's CL is much easier to merge but would give us a filemap library that will be a small subset of libutils, which would still contain the filemap support. In general this doesn't seem like a great idea. Does it potentially get us into trouble with dynamic library dependencies pulling in both?

The one disadvantage that I'm aware of is the potential waste of memory. If there are others I very much like to know.

The motivation for doing this is to reduce the surface on /system/core libs from /art for the ART APEX module in the Mainline project. One approach for doing that is to statically link several of those libs into the ART shared libs, so it's not just about this particular case. I'm working on it for libs that don't expose an ABI over I/O etc (e.g. liblog will not be statically linked for that reason).

I agree it isn't very pretty to link libs statically into multiple dynamic libs, but the alternative as things stand is to ensure ABI compatibility for all possible combinations of APEX modules that may appear on a device.

Patch-set: 9