Update patch set 1
Patch Set 1:

Issue is that apexd spawns a Binder thread that reports a comm of Binder:743_1, the main thread reports a comm of apexd. Both report a cmdline of /system/bin/apexd.

We should consider altering the BlackList Code in llkd to also match to the android::base:Basename of the cmdline, rather than just on the full name of the cmdline. This will solve the problem, and may lead to additional simplification of the blacklist.

I will push a change to _this_ CL once I have tested the solution.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>