Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> is that true? the docs say _parent_ directories. and when i added
> bionic/linker/TEST_MAPPING and ran `atest --info` in bionic/ it
> claimed there were no tests.
> 
> ah, thanks. probably worth making one of the examples in the docs show
> that this is possible.

Elliott,

For your test mapping file being in bionic/linker/TEST_MAPPING, you need to run `atest --info` in bionic/linker/ or any of its sub directories to be able to load the tests in that TEST_MAPPING file. atest searches test mapping files in the parent directories.

Or you can use --include-subdir option. That allows atest command to search for tests in the sub-directories as well, not just the parent dirs.
usage of this option is described in go/test-mapping
https://docs.google.com/document/d/1Q-RM3xs4CETiFaVh6S-kzOPLc42G85ivsXmWSgfLdFQ/edit?ts=5b48e58e#bookmark=id.paqgq7whkzf2

Patch-set: 2